refactor: 매장 주문 생성 정책 인터페이스 분리
