+++ src/test/java/kitchenpos/eatinorders/feedback/domain/EatInOrderTest.java
@@ -1,10 +1,5 @@
 package kitchenpos.eatinorders.feedback.domain;
 
-import kitchenpos.eatinorders.feedback.InMemoryOrderTableRepository;
-import kitchenpos.menus.tobe.InMemoryMenuRepository;
-import kitchenpos.menus.tobe.domain.Menu;
-import kitchenpos.menus.tobe.domain.MenuProduct;
-import kitchenpos.menus.tobe.domain.MenuRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -16,30 +11,17 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class EatInOrderTest {
-    private OrderTableRepository orderTableRepository = new InMemoryOrderTableRepository();
-    private MenuRepository menuRepository = new InMemoryMenuRepository();
-    private OrderCreatePolicy orderCreatePolicy = new OrderCreatePolicy(orderTableRepository, menuRepository);
-
     private List<OrderLineItem> orderLineItems;
-    private OrderTable orderTable;
-    private Menu menu;
 
     @BeforeEach
     void setUp() {
         orderLineItems = List.of(new OrderLineItem(1L, BigDecimal.valueOf(20000), 2));
-
-        orderTable = new OrderTable(1L, "1번 테이블");
-        orderTableRepository.save(orderTable);
-        orderTable.use();
-
-        menu = new Menu(1L, "후라이드 치킨", 20000, 1L, List.of(new MenuProduct(1L, 1)));
-        menuRepository.save(menu);
     }
 
     @DisplayName("매장 주문을 생성한다.")
     @Test
     void create() {
-        EatInOrder eatInOrder = new EatInOrder(orderLineItems, 1L, orderCreatePolicy);
+        EatInOrder eatInOrder = new EatInOrder(orderLineItems, 1L, new FakeOrderCreatePolicy());
 
         assertThat(eatInOrder.getOrderStatus()).isEqualTo(OrderStatus.WAITING);
     }
@@ -47,7 +29,7 @@ class EatInOrderTest {
     @DisplayName("매장 주문을 접수한다.")
     @Test
     void accept() {
-        EatInOrder eatInOrder = new EatInOrder(orderLineItems, 1L, orderCreatePolicy);
+        EatInOrder eatInOrder = new EatInOrder(orderLineItems, 1L, new FakeOrderCreatePolicy());
 
         eatInOrder.accept();
 
@@ -57,7 +39,7 @@ class EatInOrderTest {
     @DisplayName("WAITING 상태가 아닌 매장 주문은 접수할 수 없다.")
     @Test
     void acceptFail() {
-        EatInOrder eatInOrder = new EatInOrder(orderLineItems, 1L, orderCreatePolicy);
+        EatInOrder eatInOrder = new EatInOrder(orderLineItems, 1L, new FakeOrderCreatePolicy());
         eatInOrder.accept();
 
         assertThatThrownBy(eatInOrder::accept)
@@ -67,7 +49,7 @@ class EatInOrderTest {
     @DisplayName("매장 주문을 서빙한다.")
     @Test
     void serve() {
-        EatInOrder eatInOrder = new EatInOrder(orderLineItems, 1L, orderCreatePolicy);
+        EatInOrder eatInOrder = new EatInOrder(orderLineItems, 1L, new FakeOrderCreatePolicy());
         eatInOrder.accept();
 
         eatInOrder.serve();
@@ -78,7 +60,7 @@ class EatInOrderTest {
     @DisplayName("ACCEPT 상태가 아닌 매장 주문은 서빙할 수 없다.")
     @Test
     void serveFail() {
-        EatInOrder eatInOrder = new EatInOrder(orderLineItems, 1L, orderCreatePolicy);
+        EatInOrder eatInOrder = new EatInOrder(orderLineItems, 1L, new FakeOrderCreatePolicy());
 
         assertThatThrownBy(eatInOrder::serve)
                 .isInstanceOf(IllegalStateException.class);
@@ -87,7 +69,7 @@ class EatInOrderTest {
     @DisplayName("매장 주문을 완료한다.")
     @Test
     void complete() {
-        EatInOrder eatInOrder = new EatInOrder(orderLineItems, 1L, orderCreatePolicy);
+        EatInOrder eatInOrder = new EatInOrder(orderLineItems, 1L, new FakeOrderCreatePolicy());
         eatInOrder.accept();
         eatInOrder.serve();
 
@@ -99,9 +81,16 @@ class EatInOrderTest {
     @DisplayName("SERVED 상태가 아닌 매장 주문은 완료할 수 없다.")
     @Test
     void completeFail() {
-        EatInOrder eatInOrder = new EatInOrder(orderLineItems, 1L, orderCreatePolicy);
+        EatInOrder eatInOrder = new EatInOrder(orderLineItems, 1L, new FakeOrderCreatePolicy());
 
         assertThatThrownBy(eatInOrder::complete)
                 .isInstanceOf(IllegalStateException.class);
+    }
+
+    static class FakeOrderCreatePolicy implements OrderCreatePolicy {
+        @Override
+        public void validate(List<OrderLineItem> orderLineItems, Long orderTableId) {
+
+        }
     }
 }
